create card module, move card-status and card to card module
