+++ alcs-api/src/application/application-code/application-code.controller.ts
@@ -6,8 +6,8 @@ import * as config from 'config';
 import { RoleGuard } from '../../common/authorization/role.guard';
 import { ANY_AUTH_ROLE } from '../../common/authorization/roles';
 import { UserRoles } from '../../common/authorization/roles.decorator';
-import { CardStatusDto } from '../application-status/card-status.dto';
-import { CardStatus } from '../application-status/card-status.entity';
+import { CardStatusDto } from '../../card/card-status/card-status.dto';
+import { CardStatus } from '../../card/card-status/card-status.entity';
 import { ApplicationMasterCodesDto } from './application-code.dto';
 import { ApplicationCodeService } from './application-code.service';
 import { ApplicationRegionDto } from './application-region/application-region.dto';
