+++ alcs-api/src/application/application-code/application-code.dto.ts
@@ -1,4 +1,4 @@
-import { CardStatusDto } from '../application-status/card-status.dto';
+import { CardStatusDto } from '../../card/card-status/card-status.dto';
 import { ApplicationRegionDto } from './application-region/application-region.dto';
 import { ApplicationTypeDto } from './application-type/application-type.dto';
