+++ alcs-api/src/application/application.dto.ts
@@ -12,7 +12,7 @@ import { UserDto } from '../user/user.dto';
 import { ApplicationRegionDto } from './application-code/application-region/application-region.dto';
 import { ApplicationTypeDto } from './application-code/application-type/application-type.dto';
 import { ApplicationDecisionMeetingDto } from './application-decision-meeting/application-decision-meeting.dto';
-import { CardStatusDto } from './application-status/card-status.dto';
+import { CardStatusDto } from '../card/card-status/card-status.dto';
 
 export class CreateApplicationDto {
   @AutoMap()
