+++ alcs-api/src/board/board-status.entity.ts
@@ -1,6 +1,6 @@
 import { AutoMap } from '@automapper/classes';
 import { Column, Entity, ManyToOne, Unique } from 'typeorm';
-import { CardStatus } from '../application/application-status/card-status.entity';
+import { CardStatus } from '../card/card-status/card-status.entity';
 import { Base } from '../common/entities/base.entity';
 import { Board } from './board.entity';
