+++ alcs-api/src/card/card-status/card-status.controller.spec.ts
@@ -3,9 +3,9 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { getRepositoryToken } from '@nestjs/typeorm';
 import { initCardStatusMockEntity } from '../../common/utils/test-helpers/mockEntities';
 import { repositoryMockFactory } from '../../common/utils/test-helpers/mockTypes';
-import { ApplicationTimeTrackingService } from '../application-time-tracking.service';
-import { Application } from '../application.entity';
-import { ApplicationService } from '../application.service';
+import { ApplicationTimeTrackingService } from '../../application/application-time-tracking.service';
+import { Application } from '../../application/application.entity';
+import { ApplicationService } from '../../application/application.service';
 import { CardStatusController } from './card-status.controller';
 import { CardStatusDto } from './card-status.dto';
 import { CardStatus } from './card-status.entity';
